# fs: improve `ExistsSync` performance on Windows