@anonrig 2024-06-21
> Have you run the benchmarks under the benchmark folder? It’s not very reliable to benchmark a single fs call within a script using hyperfine since Node.js itself can take a lot more time to start up than doing that one fs call. (Perhaps around 40ms of that 48-47ms can be Node.js startup)

I agree. Unfortunately, I couldn't figure out some weird `npm` issue on my Windows machine, and couldn't make a progress due to the slowness of the benchmarks (even if I decreased the `n`).