@anonrig 2024-06-21
I've opened an issue to keep track of the Windows issue I'm facing: https://github.com/nodejs/node/issues/53538