@joyeecheung 2024-06-22
> I agree. Unfortunately, I couldn't figure out some weird npm issue on my Windows machine, and couldn't make a progress due to the slowness of the benchmarks (even if I decreased the n).

Why would you need npm access to run the benchmarks? I think they can simply be run as

```
out/Release/node benchmark/compare.js --old ./node_main --new out/Release/node --filter existsSync fs > fs.csv
```

?