@joyeecheung 2024-06-22
Oh I see in https://github.com/nodejs/node/issues/53538 you were trying to run `node-benchmark-compare`. You could just send that over to any machine that can run `node-benchmark-compare` to analyze the data, the csv is portable.