@anonrig 2024-06-23
> Microbenchmark CI https://ci.nodejs.org/view/Node.js%20benchmark/job/benchmark-node-micro-benchmarks/1571/

It seems, this PR is 15% faster