@BridgeAR 2024-06-24
> It seems, this PR is 15% faster

That only applies to the non-existing part. The question for me is what is more likely: existing or non-existing?

```
                                                          confidence improvement accuracy (*)   (**)  (***)
fs/bench-existsSync.js n=1000000 type='existing'                 ***     -0.92 %       ±0.29% ±0.39% ±0.51%
fs/bench-existsSync.js n=1000000 type='non-existing'             ***     15.26 %       ±1.16% ±1.55% ±2.04%
fs/bench-existsSync.js n=1000000 type='non-flat-existing'        ***      2.50 %       ±0.39% ±0.53% ±0.69%
```