@anonrig 2024-06-24
> That only applies to the non-existing part. The question for me is what is more likely: existing or non-existing?

@BridgeAR for unix, the difference is 15% but on Windows, but on windows, it reduces the number of OS calls.