@anonrig 2024-06-24
> Did you test this hypothesis, or is it guaranteed by `std::filesystem::exists`?

I was referencing the cppreference (https://en.cppreference.com/w/cpp/filesystem/exists)