@tniessen 2024-06-24
> > Did you test this hypothesis, or is it guaranteed by `std::filesystem::exists`?
> 
> I was referencing the cppreference (https://en.cppreference.com/w/cpp/filesystem/exists)

I don't see anything in that reference that would guarantee that the standard library function makes a single system call (or the Windows equivalent) and not multiple system calls.